+++ src/main/kotlin/org/pcsoft/framework/pluggiat/manifest/IconFormatException.kt
@@ -1,0 +1,6 @@
+package org.pcsoft.framework.pluggiat.manifest
+
+/**
+ * Thrown when the image format of a plugin's `icon` field cannot be recognised.
+ */
+internal class IconFormatException(message: String) : Exception(message)